- Not allowed order cancelation via payment window for the order that is paid.

+++ Controller/Standard/Cancel.php
@@ -55,6 +55,11 @@ class Cancel extends \Magento\Framework\App\Action\Action
     protected $resultJsonFactory;
 
     /**
+     * @var \Magento\Sales\Api\Data\TransactionSearchResultInterfaceFactory
+     */
+    protected $transactionSearchResultInterfaceFactory;
+
+    /**
      * Constructor
      *
      * @param \Magento\Framework\App\Action\Context $context
@@ -67,6 +72,7 @@ class Cancel extends \Magento\Framework\App\Action\Action
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
      * @param \Magento\Framework\Controller\Result\JsonFactory $resultJsonFactory
      * @param \Radarsofthouse\Reepay\Helper\Logger $logger
+     * @param \Magento\Sales\Api\Data\TransactionSearchResultInterfaceFactory $transactionSearchResultInterfaceFactory
      */
     public function __construct(
         \Magento\Framework\App\Action\Context $context,
@@ -78,7 +84,8 @@ class Cancel extends \Magento\Framework\App\Action\Action
         \Radarsofthouse\Reepay\Helper\Data $reepayHelper,
         \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
         \Magento\Framework\Controller\Result\JsonFactory $resultJsonFactory,
-        \Radarsofthouse\Reepay\Helper\Logger $logger
+        \Radarsofthouse\Reepay\Helper\Logger $logger,
+        \Magento\Sales\Api\Data\TransactionSearchResultInterfaceFactory $transactionSearchResultInterfaceFactory
     ) {
         $this->resultPageFactory = $resultPageFactory;
         $this->request = $request;
@@ -90,6 +97,7 @@ class Cancel extends \Magento\Framework\App\Action\Action
         $this->scopeConfig = $scopeConfig;
         $this->resultJsonFactory = $resultJsonFactory;
         $this->logger = $logger;
+        $this->transactionSearchResultInterfaceFactory = $transactionSearchResultInterfaceFactory;
 
         parent::__construct($context);
     }
@@ -122,23 +130,31 @@ class Cancel extends \Magento\Framework\App\Action\Action
         $cancelConfig = $this->reepayHelper->getConfig('cancel_order_after_payment_cancel', $order->getStoreId());
 
         if ($cancelConfig && $order->canCancel()) {
-            $order->cancel();
-            $order->addStatusHistoryComment('Reepay : order have been cancelled by payment page');
-            $order->save();
-            $this->logger->addDebug('Cancelled order : ' . $orderId);
-            $apiKey = $this->reepayHelper->getApiKey($order->getStoreId());
-            $payment = $order->getPayment();
-            $this->reepayHelper->setReepayPaymentState($payment, 'cancelled');
-            // delete reepay session
-            $sessionRes = $this->reepaySession->delete(
-                $apiKey,
-                $id
-            );
-            $this->checkoutSession->restoreQuote();
-            $this->checkoutSession->unsLastQuoteId()
-                ->unsLastSuccessQuoteId()
-                ->unsLastOrderId()
-                ->unsLastRealOrderId();
+
+            $transactions = $this->transactionSearchResultInterfaceFactory->create()->addOrderIdFilter($order->getId());
+
+            // don't allowed the cancelation if already have transactions (payment is paid)
+            if( count($transactions->getItems()) == 0 ){
+                $order->cancel();
+                $order->addStatusHistoryComment('Reepay : order have been cancelled by payment page');
+                $order->save();
+                $this->logger->addDebug('Cancelled order : ' . $orderId);
+                $apiKey = $this->reepayHelper->getApiKey($order->getStoreId());
+                $payment = $order->getPayment();
+                $this->reepayHelper->setReepayPaymentState($payment, 'cancelled');
+                // delete reepay session
+                $sessionRes = $this->reepaySession->delete(
+                    $apiKey,
+                    $id
+                );
+                $this->checkoutSession->restoreQuote();
+                $this->checkoutSession->unsLastQuoteId()
+                    ->unsLastSuccessQuoteId()
+                    ->unsLastOrderId()
+                    ->unsLastRealOrderId();
+            }else{
+                $this->logger->addDebug('The payment is done : ignore cancellation for order ' . $orderId);
+            }
         }
 
         // unset reepay session id on checkout session
